deployments/docker-compose/arizephoenix/proxy/requirements.txt
fastapi
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
httpx
requests
